refactor behaviors to their own files
