+++ src/player/resources/character.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 use serde::Deserialize;
 use std::collections::HashMap;
 
-use crate::spawnable::{Health, ProjectileType};
+use crate::{misc::Health, spawnable::ProjectileType};
 
 /// Contains data necessary to create a player entity.
 /// A character is chosen at the beginning of the game.
