+++ src/spawnable/mod.rs
@@ -12,8 +12,7 @@ mod projectile;
 mod spawn;
 
 pub use self::mob::{
-    health::Health, mob_execute_behavior_system, spawn_mob, MobBehavior, MobComponent, MobData,
-    MobsResource,
+    mob_execute_behavior_system, spawn_mob, MobBehavior, MobComponent, MobData, MobsResource,
 };
 pub use self::projectile::{
     projectile_execute_behavior_system, spawn_projectile, ProjectileData, ProjectileResource,
